tutorials/3-Practical/streamlit-leaderboard/requirements.txt
streamlit>=0.62.0
argon2-cffi>=20.1.0
pandas>=1.0.5
bokeh>=2.1.1
numpy>=1.18.5
scikit-learn>=0.23.1
